Superstore_web/requirements.txt
asgiref==3.5.0
Django==4.0.4
django-environ==0.9.0
djangorestframework==3.14.0
mysqlclient==2.1.1
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.2
